# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trips/{id}": {
      "get": {
        "operationId": "TripsController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find trip by id."
          }
        },
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "operationId": "TripsController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update trip."
          }
        },
        "tags": [
          "trips"
        ]
      },
      "delete": {
        "operationId": "TripsController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete trip."
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
...
    "/trip-requests": {
      "get": {
        "operationId": "TripRequestsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List pending traveler trip requests."
          }
        },
        "tags": [
          "trip-requests"
        ]
      },
      "post": {
        "operationId": "TripRequestsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create traveler trip request."
          }
        },
        "tags": [
          "trip-requests"
        ]
      }
    },
    "/trip-requests/{id}/accept": {
      "post": {
        "operationId": "TripRequestsController_accept",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Travel partner accepts a traveler request."
          }
        },
        "tags": [
          "trip-requests"
        ]
      }
    },
    "/trip-requests/{id}": {
      "delete": {
        "operationId": "TripRequestsController_cancel",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancel or delete a traveler trip request."
          }
        },
        "tags": [
          "trip-requests"
        ]
      }
    },
...
    "/messages": {
      "get": {
        "operationId": "MessagesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "tripId",
            "required": false,
            "in": "query",
            "description": "Filter messages by trip id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "Filter messages for user email.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-email",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List messages."
          }
        },
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Send a message and notify recipients."
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/{id}": {
      "get": {
        "operationId": "MessagesController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      },
      "patch": {
        "operationId": "MessagesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      },
      "delete": {
        "operationId": "MessagesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter notifications for a role key.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "Filter notifications for user email.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-email",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List notifications."
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mark notification as read for a role."
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dream Destination Review 4 API",
    "description": "NestJS in-memory REST API with RBAC header authorization for traveler, travel partner, tour guide, vendor, support, and super user portals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required on every API. Use Super User, Traveler, Travel Partner, Tour Guide, Vendor, or Support Executive."
      }
    },
    "schemas": {
...
      "UpdateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Goa Beach Tour"
          },
          "destination": {
            "type": "string",
            "example": "Goa, India"
          },
          "travelerName": {
            "type": "string",
            "example": "Rahul Sharma"
          },
          "travelerEmail": {
            "type": "string",
            "example": "rahul@example.com"
          },
          "startDate": {
            "type": "string",
            "example": "2026-05-12"
          },
          "endDate": {
            "type": "string",
            "example": "2026-05-16"
          },
          "adults": {
            "type": "number",
            "example": 2
          },
          "children": {
            "type": "number",
            "example": 0
          },
          "budget": {
            "type": "number",
            "example": 4500
          },
          "status": {
            "type": "string",
            "example": "planning"
          },
          "interests": {
            "example": [
              "beach",
              "heritage"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "Need family-friendly activities."
          }
        }
      },
...
      "CreateTripRequestDto": {
        "type": "object",
        "properties": {
          "travelerName": {
            "type": "string",
            "example": "Sarah Johnson"
          },
          "travelerEmail": {
            "type": "string",
            "example": "traveler@gmail.com"
          },
          "packageName": {
            "type": "string",
            "example": "Maldives Escape"
          },
          "destination": {
            "type": "string",
            "example": "Male, Maldives"
          },
          "startDate": {
            "type": "string",
            "example": "2026-06-10"
          },
          "endDate": {
            "type": "string",
            "example": "2026-06-16"
          },
          "adults": {
            "type": "number",
            "example": 2
          },
          "children": {
            "type": "number",
            "example": 0
          },
          "budget": {
            "type": "number",
            "example": 5200
          },
          "interests": {
            "example": [
              "beach",
              "snorkeling"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "Ocean-view resort preferred."
          }
        },
        "required": [
          "travelerName",
          "packageName",
          "destination"
        ]
      },
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "TRIP-101"
          },
          "fromRole": {
            "type": "string",
            "example": "traveler"
          },
          "fromName": {
            "type": "string",
            "example": "Rahul Sharma"
          },
          "toRoles": {
            "example": [
              "guide",
              "vendor"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "body": {
            "type": "string",
            "example": "What time is the pickup?"
          }
        },
        "required": [
          "tripId",
          "fromRole",
          "fromName",
          "body"
        ]
      },
      "UpdateMessageDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "TRIP-101"
          },
          "fromRole": {
            "type": "string",
            "example": "traveler"
          },
          "fromName": {
            "type": "string",
            "example": "Rahul Sharma"
          },
          "toRoles": {
            "example": [
              "guide",
              "vendor"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "body": {
            "type": "string",
            "example": "What time is the pickup?"
          }
        }
      },
      "ReadNotificationDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "traveler"
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}